+++ trading_vision/__init__.py
@@ -1,3 +1,3 @@
 """Trading Vision application package."""
 
-__version__ = "0.60.0"
+__version__ = "0.61.0"